examples/rag_example/requirements.txt
requests
beautifulsoup4
google-generativeai
faiss-cpu
numpy
pyyaml